authentication improved

+++ main/views.py
@@ -1,40 +1,52 @@
-from django.http import HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404, redirect, render
 from .models import *
 from .forms import *
 from django.views.generic import *
+from django.contrib.auth.decorators import login_required
+
+
+
 
 def homeView(request):
-    if request.method == 'POST':
-        form = ToDoListForm(request.POST)
-        if form.is_valid():
-            cd = form.cleaned_data
-            ToDoItem.objects.create(text=cd['text'], author=request.user)
-            return HttpResponseRedirect("/")
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = ToDoListForm(request.POST)
+            if form.is_valid():
+                cd = form.cleaned_data
+                ToDoItem.objects.create(text=cd['text'], author=request.user)
+                return HttpResponseRedirect("/")
+        else:
+            return render(request, 'main.html', {'ToDoItem':ToDoItem.objects.filter(author=request.user), 'ToDoListForm':ToDoListForm()})
     else:
-        return render(request, 'main.html', {'ToDoItem':ToDoItem.objects.all(), 'ToDoListForm':ToDoListForm()})
+        return render(request, 'main.html', {})
 
 
 def deleteView(request, id):
-    obj = get_object_or_404(ToDoItem, id = id)
-    if request.method == 'POST':
-        obj.delete()
-        return HttpResponseRedirect("/")
-    
-    return render(request, 'delete.html',{'ToDoItem':ToDoItem.objects.get(id=id)})
-
+    if request.user==ToDoItem.objects.get(id=id).author:
+        obj = get_object_or_404(ToDoItem, id = id)
+        if request.method == 'POST':
+            obj.delete()
+            return HttpResponseRedirect("/")
+        
+        return render(request, 'delete.html',{'ToDoItem':ToDoItem.objects.get(id=id)})
+    else:
+        return HttpResponse("You are not elligible to delete this item")
 
 def updateView(request, id):
-    obj = ToDoItem.objects.get(id=id)
-    if request.method == 'POST':
-        form = ToDoListForm(request.POST, instance = obj)
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect("/")
+    if request.user==ToDoItem.objects.get(id=id).author:
+        obj = ToDoItem.objects.get(id=id)
+        if request.method == 'POST':
+            form = ToDoListForm(request.POST, instance = obj)
+            if form.is_valid():
+                form.save()
+                return HttpResponseRedirect("/")
+        else:
+            return render(request, 'update.html', {'ToDoListForm':ToDoListForm(instance=obj)})
+        
+        return render(request, 'update.html', {})
     else:
-        return render(request, 'update.html', {'ToDoListForm':ToDoListForm(instance=obj)})
-    
-    return render(request, 'update.html', {})
+        return HttpResponse("You are not elligible to edit this item")
